chore: rename terradraw text mode file names to obey Terradraw's naming convention
