+++ src/lib/controls/MaplibreTerradrawControl.ts
@@ -26,10 +26,7 @@ import type {
 } from '../interfaces';
 import { defaultControlOptions, getDefaultModeOptions } from '../constants';
 import { capitalize, cleanMaplibreStyle, TERRADRAW_SOURCE_IDS, ModalDialog } from '../helpers';
-import type {
-	MaplibreTerradrawTextMode,
-	TextModeStyling
-} from '../modes/MaplibreTerradrawTextMode';
+import type { MaplibreTerradrawTextMode, TextModeStyling } from '../modes/TerradrawTextMode';
 
 /**
  * Maplibre GL Terra Draw Control
