+++ src/lib/modes/TerradrawTextMode.ts
